REFACTOR: Extract shared pole figure chrome into PoleFigureChrome module
